+++ src/_NavigableApp/App.js
@@ -9,13 +9,16 @@ import Tab from 'react-bootstrap/Tab'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 
+//import Nav from './PrototypeNav';
+//import Footer from './PrototypeFooter';
+
 import PlotlyDemo from '../_Dashboards/PlotlyDemo/Dashboard.js';
 import RechartsDemo from '../_Dashboards/RechartsDemo/Dashboard.js';
 import UsersMostRetweeted from '../_Dashboards/UsersMostRetweeted/Dashboard.js';
 import StatusesMostRetweeted from '../_Dashboards/StatusesMostRetweeted/Dashboard.js';
 import TopProfileTags from '../_Dashboards/TopProfileTags/Dashboard.js';
 import TopStatusTags from '../_Dashboards/TopStatusTags/Dashboard.js';
-//import UserOpinion from '../_Dashboards/UserOpinion/Dashboard.js';
+import UserOpinion from '../_Dashboards/UserOpinion/Dashboard.js';
 
 // TODO
 const Home = () => <h2>Home</h2>
@@ -42,6 +45,7 @@ const App = () => (
                 <Nav.Item><Nav.Link eventKey="top_status_tags">Bot Tweet Hashtags</Nav.Link></Nav.Item>
                 <Nav.Item><Nav.Link eventKey="top_statuses">Bot Retweets</Nav.Link></Nav.Item>
                 <Nav.Item><Nav.Link eventKey="top_users">Bot Retweeted Users</Nav.Link></Nav.Item>
+                <Nav.Item><Nav.Link eventKey="user_opinion">User Opinions</Nav.Link></Nav.Item>
               </Nav>
             </Col>
             <Col sm={9}>
@@ -54,6 +58,7 @@ const App = () => (
                 <Tab.Pane eventKey="top_status_tags"><TopStatusTags/></Tab.Pane>
                 <Tab.Pane eventKey="top_statuses"><StatusesMostRetweeted/></Tab.Pane>
                 <Tab.Pane eventKey="top_users"><UsersMostRetweeted/></Tab.Pane>
+                <Tab.Pane eventKey="user_opinion"><UserOpinion/></Tab.Pane>
               </Tab.Content>
             </Col>
           </Row>
